+++ src/main/java/rusty/vanillo/registry/VBlocks.java
@@ -8,7 +8,6 @@ import net.minecraft.world.level.block.StairBlock;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
 import net.minecraft.world.level.material.Material;
-import net.minecraftforge.common.ToolType;
 import net.minecraftforge.fmllegacy.RegistryObject;
 import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
@@ -26,9 +25,8 @@ public final class VBlocks {
     private static final BlockBehaviour.StatePredicate NEVER = (s, e, x) -> false;
 
     // Extra slab blocks
-    public static final RegistryObject<SlabBlock> DIRT_SLAB = registerCustomSlab("dirt_slab", () -> BlockBehaviour.Properties.copy(Blocks.DIRT).harvestTool(ToolType.SHOVEL));
-    public static final RegistryObject<SlabBlock> COARSE_DIRT_SLAB = registerCustomSlab("coarse_dirt_slab", () -> BlockBehaviour.Properties.copy(Blocks.COARSE_DIRT).harvestTool(ToolType.SHOVEL));
-    //public static final RegistryObject<VGrassSlab> GRASS_SLAB = BLOCKS.register("grass_slab", () -> new VGrassSlab(AbstractBlock.Properties.of(Material.GRASS).randomTicks().strength(0.6F).sound(SoundType.GRASS).noOcclusion()));
+    public static final RegistryObject<SlabBlock> DIRT_SLAB = registerCustomSlab("dirt_slab", () -> BlockBehaviour.Properties.copy(Blocks.DIRT));
+    public static final RegistryObject<SlabBlock> COARSE_DIRT_SLAB = registerCustomSlab("coarse_dirt_slab", () -> BlockBehaviour.Properties.copy(Blocks.COARSE_DIRT));
 
     // Stone Brick Bricks + Brick Bricks
     public static final RegistryObject<Block> BRICK_BRICKS = BLOCKS.register("brick_bricks", () -> new Block(BlockBehaviour.Properties.copy(Blocks.BRICKS)));
@@ -39,18 +37,18 @@ public final class VBlocks {
     public static final RegistryObject<StairBlock> STONE_BRICK_BRICK_STAIRS = registerStairs("stone_brick_brick_stairs", STONE_BRICK_BRICKS);
 
     // Recycler
-    public static final RegistryObject<Block> RECYCLER = BLOCKS.register("recycler", () -> new RecyclerBlock(BlockBehaviour.Properties.of(Material.HEAVY_METAL).sound(SoundType.ANVIL).requiresCorrectToolForDrops().harvestTool(ToolType.PICKAXE).lightLevel(state -> state.getValue(BlockStateProperties.LIT) ? 13 : 0)));
+    public static final RegistryObject<Block> RECYCLER = BLOCKS.register("recycler", () -> new RecyclerBlock(BlockBehaviour.Properties.of(Material.HEAVY_METAL).sound(SoundType.ANVIL).strength(5.0f).requiresCorrectToolForDrops().lightLevel(state -> state.getValue(BlockStateProperties.LIT) ? 13 : 0)));
 
     // Rails
-    public static final RegistryObject<Block> WOODEN_RAIL = BLOCKS.register("wooden_rail", () -> new VRailBlock(0.1f, BlockBehaviour.Properties.of(Material.DECORATION).harvestTool(ToolType.AXE).sound(SoundType.LADDER).noCollission().strength(0.7F)));
-    public static final RegistryObject<Block> GLOWSTONE_RAIL = BLOCKS.register("glowstone_rail", () -> new VRailBlock(0.6f, BlockBehaviour.Properties.of(Material.DECORATION).harvestTool(ToolType.PICKAXE).sound(SoundType.GLASS).noCollission().strength(0.7F).lightLevel(blockState -> 14)));
-    public static final RegistryObject<Block> DIAMOND_POWERED_RAIL = BLOCKS.register("diamond_powered_rail", () -> new VPoweredRailBlock(0.75f, BlockBehaviour.Properties.of(Material.DECORATION).harvestTool(ToolType.PICKAXE).sound(SoundType.METAL).noCollission().strength(0.7F)));
-    public static final RegistryObject<Block> NETHERITE_POWERED_RAIL = BLOCKS.register("netherite_powered_rail", () -> new VPoweredRailBlock(1f, BlockBehaviour.Properties.of(Material.DECORATION).harvestTool(ToolType.PICKAXE).sound(SoundType.NETHERITE_BLOCK).noCollission().strength(0.7F)));
-    public static final RegistryObject<Block> VOID_POWERED_RAIL = BLOCKS.register("void_powered_rail", () -> new VPoweredRailBlock(2.5f, BlockBehaviour.Properties.of(Material.DECORATION).harvestTool(ToolType.PICKAXE).sound(SoundType.NETHERITE_BLOCK).noCollission().strength(0.7F).lightLevel(blockState -> 9)));
+    public static final RegistryObject<Block> WOODEN_RAIL = BLOCKS.register("wooden_rail", () -> new VRailBlock(0.1f, BlockBehaviour.Properties.of(Material.DECORATION).sound(SoundType.LADDER).noCollission().strength(0.7F)));
+    public static final RegistryObject<Block> GLOWSTONE_RAIL = BLOCKS.register("glowstone_rail", () -> new VRailBlock(0.6f, BlockBehaviour.Properties.of(Material.DECORATION).sound(SoundType.GLASS).noCollission().strength(0.7F).lightLevel(blockState -> 14)));
+    public static final RegistryObject<Block> DIAMOND_POWERED_RAIL = BLOCKS.register("diamond_powered_rail", () -> new VPoweredRailBlock(0.75f, BlockBehaviour.Properties.of(Material.DECORATION).sound(SoundType.METAL).noCollission().strength(0.7F)));
+    public static final RegistryObject<Block> NETHERITE_POWERED_RAIL = BLOCKS.register("netherite_powered_rail", () -> new VPoweredRailBlock(1f, BlockBehaviour.Properties.of(Material.DECORATION).sound(SoundType.NETHERITE_BLOCK).noCollission().strength(0.7F)));
+    public static final RegistryObject<Block> VOID_POWERED_RAIL = BLOCKS.register("void_powered_rail", () -> new VPoweredRailBlock(2.5f, BlockBehaviour.Properties.of(Material.DECORATION).sound(SoundType.NETHERITE_BLOCK).noCollission().strength(0.7F).lightLevel(blockState -> 9)));
 
     // Void Equipment
-    public static final RegistryObject<Block> VOID_ORE = BLOCKS.register("void_ore", () -> new VoidOre(BlockBehaviour.Properties.of(Material.METAL).sound(SoundType.ANCIENT_DEBRIS).harvestTool(ToolType.PICKAXE).harvestLevel(4).requiresCorrectToolForDrops().strength(50.0F, 1200.0F).lightLevel(blockState -> 13)));
-    public static final RegistryObject<Block> VOID_BLOCK = BLOCKS.register("void_block", () -> new Block(BlockBehaviour.Properties.of(Material.METAL).sound(SoundType.NETHERITE_BLOCK).harvestTool(ToolType.PICKAXE).harvestLevel(4).requiresCorrectToolForDrops().strength(50.0F, 1200.0F).lightLevel(blockState -> 9).noOcclusion().isViewBlocking(NEVER)));
+    public static final RegistryObject<Block> VOID_ORE = BLOCKS.register("void_ore", () -> new VoidOre(BlockBehaviour.Properties.of(Material.METAL).sound(SoundType.ANCIENT_DEBRIS).requiresCorrectToolForDrops().strength(50.0F, 1200.0F).lightLevel(blockState -> 13)));
+    public static final RegistryObject<Block> VOID_BLOCK = BLOCKS.register("void_block", () -> new Block(BlockBehaviour.Properties.of(Material.METAL).sound(SoundType.NETHERITE_BLOCK).requiresCorrectToolForDrops().strength(50.0F, 1200.0F).lightLevel(blockState -> 9).noOcclusion().isViewBlocking(NEVER)));
 
     public static RegistryObject<SlabBlock> registerSlab(String name, Supplier<Block> block) {
         return registerCustomSlab(name, () -> BlockBehaviour.Properties.copy(block.get()));